third_party/WebKit/LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/idlharness-expected.txt

idlharness test

This test validates the WebIDL included in the Navigation Timing specification.

interface PerformanceTiming {
  readonly attribute unsigned long long navigationStart;
  readonly attribute unsigned long long unloadEventStart;
  readonly attribute unsigned long long unloadEventEnd;
  readonly attribute unsigned long long redirectStart;
  readonly attribute unsigned long long redirectEnd;
  readonly attribute unsigned long long fetchStart;
  readonly attribute unsigned long long domainLookupStart;
  readonly attribute unsigned long long domainLookupEnd;
  readonly attribute unsigned long long connectStart;
  readonly attribute unsigned long long connectEnd;
  readonly attribute unsigned long long secureConnectionStart;
  readonly attribute unsigned long long requestStart;
  readonly attribute unsigned long long responseStart;
  readonly attribute unsigned long long responseEnd;
  readonly attribute unsigned long long domLoading;
  readonly attribute unsigned long long domInteractive;
  readonly attribute unsigned long long domContentLoadedEventStart;
  readonly attribute unsigned long long domContentLoadedEventEnd;
  readonly attribute unsigned long long domComplete;
  readonly attribute unsigned long long loadEventStart;
  readonly attribute unsigned long long loadEventEnd;
};

interface PerformanceNavigation {
  const unsigned short TYPE_NAVIGATE = 0;
  const unsigned short TYPE_RELOAD = 1;
  const unsigned short TYPE_BACK_FORWARD = 2;
  const unsigned short TYPE_RESERVED = 255;
  readonly attribute unsigned short type;
  readonly attribute unsigned short redirectCount;
};

interface Performance {
  readonly attribute PerformanceTiming timing;
  readonly attribute PerformanceNavigation navigation;
};

partial interface Window {
  [Replaceable] readonly attribute Performance performance;
};

FAIL Window interface: attribute performance assert_true: The prototype object must have a property "performance" expected true got false
PASS Window must be primary interface of window 
FAIL Stringification of window assert_equals: class string of window expected "[object Window]" but got "[object global]"
FAIL Window interface: window must inherit property "performance" with the proper type (0) assert_inherits: property "performance" found on object expected in prototype chain
PASS PerformanceTiming interface: existence and properties of interface object 
FAIL PerformanceTiming interface: existence and properties of interface prototype object assert_equals: class string of PerformanceTiming.prototype expected "[object PerformanceTimingPrototype]" but got "[object Object]"
PASS PerformanceTiming interface: existence and properties of interface prototype object's "constructor" property 
FAIL PerformanceTiming interface: attribute navigationStart assert_true: The prototype object must have a property "navigationStart" expected true got false
FAIL PerformanceTiming interface: attribute unloadEventStart assert_true: The prototype object must have a property "unloadEventStart" expected true got false
FAIL PerformanceTiming interface: attribute unloadEventEnd assert_true: The prototype object must have a property "unloadEventEnd" expected true got false
FAIL PerformanceTiming interface: attribute redirectStart assert_true: The prototype object must have a property "redirectStart" expected true got false
FAIL PerformanceTiming interface: attribute redirectEnd assert_true: The prototype object must have a property "redirectEnd" expected true got false
FAIL PerformanceTiming interface: attribute fetchStart assert_true: The prototype object must have a property "fetchStart" expected true got false
FAIL PerformanceTiming interface: attribute domainLookupStart assert_true: The prototype object must have a property "domainLookupStart" expected true got false
FAIL PerformanceTiming interface: attribute domainLookupEnd assert_true: The prototype object must have a property "domainLookupEnd" expected true got false
FAIL PerformanceTiming interface: attribute connectStart assert_true: The prototype object must have a property "connectStart" expected true got false
FAIL PerformanceTiming interface: attribute connectEnd assert_true: The prototype object must have a property "connectEnd" expected true got false
FAIL PerformanceTiming interface: attribute secureConnectionStart assert_true: The prototype object must have a property "secureConnectionStart" expected true got false
FAIL PerformanceTiming interface: attribute requestStart assert_true: The prototype object must have a property "requestStart" expected true got false
FAIL PerformanceTiming interface: attribute responseStart assert_true: The prototype object must have a property "responseStart" expected true got false
FAIL PerformanceTiming interface: attribute responseEnd assert_true: The prototype object must have a property "responseEnd" expected true got false
FAIL PerformanceTiming interface: attribute domLoading assert_true: The prototype object must have a property "domLoading" expected true got false
FAIL PerformanceTiming interface: attribute domInteractive assert_true: The prototype object must have a property "domInteractive" expected true got false
FAIL PerformanceTiming interface: attribute domContentLoadedEventStart assert_true: The prototype object must have a property "domContentLoadedEventStart" expected true got false
FAIL PerformanceTiming interface: attribute domContentLoadedEventEnd assert_true: The prototype object must have a property "domContentLoadedEventEnd" expected true got false
FAIL PerformanceTiming interface: attribute domComplete assert_true: The prototype object must have a property "domComplete" expected true got false
FAIL PerformanceTiming interface: attribute loadEventStart assert_true: The prototype object must have a property "loadEventStart" expected true got false
FAIL PerformanceTiming interface: attribute loadEventEnd assert_true: The prototype object must have a property "loadEventEnd" expected true got false
PASS PerformanceTiming must be primary interface of window.performance.timing 
PASS Stringification of window.performance.timing 
FAIL PerformanceTiming interface: window.performance.timing must inherit property "navigationStart" with the proper type (0) assert_inherits: property "navigationStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "unloadEventStart" with the proper type (1) assert_inherits: property "unloadEventStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "unloadEventEnd" with the proper type (2) assert_inherits: property "unloadEventEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "redirectStart" with the proper type (3) assert_inherits: property "redirectStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "redirectEnd" with the proper type (4) assert_inherits: property "redirectEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "fetchStart" with the proper type (5) assert_inherits: property "fetchStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domainLookupStart" with the proper type (6) assert_inherits: property "domainLookupStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domainLookupEnd" with the proper type (7) assert_inherits: property "domainLookupEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "connectStart" with the proper type (8) assert_inherits: property "connectStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "connectEnd" with the proper type (9) assert_inherits: property "connectEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "secureConnectionStart" with the proper type (10) assert_inherits: property "secureConnectionStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "requestStart" with the proper type (11) assert_inherits: property "requestStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "responseStart" with the proper type (12) assert_inherits: property "responseStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "responseEnd" with the proper type (13) assert_inherits: property "responseEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domLoading" with the proper type (14) assert_inherits: property "domLoading" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domInteractive" with the proper type (15) assert_inherits: property "domInteractive" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domContentLoadedEventStart" with the proper type (16) assert_inherits: property "domContentLoadedEventStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domContentLoadedEventEnd" with the proper type (17) assert_inherits: property "domContentLoadedEventEnd" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "domComplete" with the proper type (18) assert_inherits: property "domComplete" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "loadEventStart" with the proper type (19) assert_inherits: property "loadEventStart" found on object expected in prototype chain
FAIL PerformanceTiming interface: window.performance.timing must inherit property "loadEventEnd" with the proper type (20) assert_inherits: property "loadEventEnd" found on object expected in prototype chain
PASS PerformanceNavigation interface: existence and properties of interface object 
FAIL PerformanceNavigation interface: existence and properties of interface prototype object assert_equals: class string of PerformanceNavigation.prototype expected "[object PerformanceNavigationPrototype]" but got "[object Object]"
PASS PerformanceNavigation interface: existence and properties of interface prototype object's "constructor" property 
PASS PerformanceNavigation interface: constant TYPE_NAVIGATE on interface object 
PASS PerformanceNavigation interface: constant TYPE_NAVIGATE on interface prototype object 
PASS PerformanceNavigation interface: constant TYPE_RELOAD on interface object 
PASS PerformanceNavigation interface: constant TYPE_RELOAD on interface prototype object 
PASS PerformanceNavigation interface: constant TYPE_BACK_FORWARD on interface object 
PASS PerformanceNavigation interface: constant TYPE_BACK_FORWARD on interface prototype object 
PASS PerformanceNavigation interface: constant TYPE_RESERVED on interface object 
PASS PerformanceNavigation interface: constant TYPE_RESERVED on interface prototype object 
FAIL PerformanceNavigation interface: attribute type assert_true: The prototype object must have a property "type" expected true got false
FAIL PerformanceNavigation interface: attribute redirectCount assert_true: The prototype object must have a property "redirectCount" expected true got false
PASS PerformanceNavigation must be primary interface of window.performance.navigation 
PASS Stringification of window.performance.navigation 
PASS PerformanceNavigation interface: window.performance.navigation must inherit property "TYPE_NAVIGATE" with the proper type (0) 
PASS PerformanceNavigation interface: window.performance.navigation must inherit property "TYPE_RELOAD" with the proper type (1) 
PASS PerformanceNavigation interface: window.performance.navigation must inherit property "TYPE_BACK_FORWARD" with the proper type (2) 
PASS PerformanceNavigation interface: window.performance.navigation must inherit property "TYPE_RESERVED" with the proper type (3) 
FAIL PerformanceNavigation interface: window.performance.navigation must inherit property "type" with the proper type (4) assert_inherits: property "type" found on object expected in prototype chain
FAIL PerformanceNavigation interface: window.performance.navigation must inherit property "redirectCount" with the proper type (5) assert_inherits: property "redirectCount" found on object expected in prototype chain
FAIL Performance interface: existence and properties of interface object assert_equals: prototype of window's property "Performance" is not Function.prototype expected function "function Empty() {}" but got function "function EventTarget() { [native code] }"
FAIL Performance interface: existence and properties of interface prototype object assert_equals: prototype of Performance.prototype is not Object.prototype expected object "[object Object]" but got object "[object Object]"
PASS Performance interface: existence and properties of interface prototype object's "constructor" property 
FAIL Performance interface: attribute timing assert_true: The prototype object must have a property "timing" expected true got false
FAIL Performance interface: attribute navigation assert_true: The prototype object must have a property "navigation" expected true got false
PASS Performance must be primary interface of window.performance 
PASS Stringification of window.performance 
FAIL Performance interface: window.performance must inherit property "timing" with the proper type (0) assert_inherits: property "timing" found on object expected in prototype chain
FAIL Performance interface: window.performance must inherit property "navigation" with the proper type (1) assert_inherits: property "navigation" found on object expected in prototype chain
